+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * [BREAKING] Added `block_to` parameter to `NodeRpcClient::sync_nullifiers` for better pagination control ([#1309](https://github.com/0xMiden/miden-client/pull/1309)).
 * [BREAKING] Removed `web-tonic` feature ([#1268](https://github.com/0xMiden/miden-client/pull/1268)).
 * [BREAKING] Updated Web Client account store functions from insert to upsert ([#1274](https://github.com/0xMiden/miden-client/pull/1274))
+* Added account creation from miden packages ([#1313](https://github.com/0xMiden/miden-client/pull/1313)).
 
 ## 0.11.6 (2025-09-18)